requirements.txt
# Core
transformers>=4.36.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# Data Sources
yfinance>=0.2.35
ccxt>=4.0.0
fredapi>=0.5.0
requests>=2.31.0

# Technical Analysis
pandas-ta>=0.3.14b0

# Time Series
prophet>=1.1.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Dashboard & Visualization
streamlit>=1.29.0
plotly>=5.18.0

# Alerts
python-telegram-bot>=20.0

# Database
sqlalchemy>=2.0.0

# Social Media Tracking
pytrends>=4.9.0

# News RSS Feeds
feedparser>=6.0.0

# Utilities
python-dotenv>=1.0.0
